Fixes in relation to PR. Squash before merge.
Makefile: Add test command for gorilla mux implementation.
plugins: Document functions and add getter for router.
server: Remove test duplication.
